requirements.txt
fastapi==0.91.0
mysqlclient==2.1.1
python-dotenv==1.0.0
SQLAlchemy==1.3.24
uvicorn==0.20.0
